requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
python-dateutil==2.9.0
